Convert conductance units

Convert conductance units

convert_conductance(.g, P = set_units(101.3246, kPa),
  R = set_units(8.31446, J/K/mol), Temp = set_units(298.15, K))

Arguments

.g
Conductance in class units. Units must convertible to one of "m/s", "umol/m^2/s/Pa", or "mol/m^2/s"

P
A pressure value of class units that is convertible to kPa. Default is 101.3246 kPa, Earth's atmospheric pressure at sea level.

R
Ideal gas constant of class units that is convertible to J/K/kg. Default is 8.31446 J/K/mol.

Temp
A temperature value of class units that is convertible to K. Default is 25 degreeC (298.15 K).

Value

@return a tibble in units "m/s", "umol/m^2/s/Pa", and "mol/m^2/s".

Examples

# library(gunit)
library(units)

#> udunits system database from /Library/Frameworks/R.framework/Versions/3.6/Resources/library/units/share/udunits

g_sc <- set_units(10, "m/s")
convert_conductance(g_sc)

#> # A tibble: 1 x 3
#>   `m/s` `umol/m^2/s/Pa` `mol/m^2/s`
#>   [m/s] [umol/m^2/Pa/s] [mol/m^2/s]
#> 1    10        4033.956     408.739

g_sc <- set_units(4, "umol/m^2/s/Pa")
convert_conductance(g_sc)

#> # A tibble: 1 x 3
#>         `m/s` `umol/m^2/s/Pa` `mol/m^2/s`
#>         [m/s] [umol/m^2/Pa/s] [mol/m^2/s]
#> 1 0.009915825               4   0.4052984

g_sc <- set_units(0.4, "mol/m^2/s")
convert_conductance(g_sc)

#> # A tibble: 1 x 3
#>         `m/s` `umol/m^2/s/Pa` `mol/m^2/s`
#>         [m/s] [umol/m^2/Pa/s] [mol/m^2/s]
#> 1 0.009786197        3.947709         0.4